+++ ApplicationCore/Interfaces/IInOutTaskService.cs
@@ -7,7 +7,7 @@ namespace ApplicationCore.Interfaces
 {
     public interface IInOutTaskService
     {
-        Task EmptyOut(int areaId,int outCount);
+        Task EmptyOut(int areaId,double outCount);
         Task EmptyEntry(string trayCode,int areaId);
 
         Task EntryApply(string fromPort,string barCode,int cargoHeight,string cargoWeight);
