+++ Web/ViewModels/OrderManager/SubOrderRowViewModel.cs
@@ -45,15 +45,15 @@ namespace Web.ViewModels.OrderManager
         /// <summary>
         /// 订单物料数量
         /// </summary>
-        public int PreCount { get; set; }
+        public double PreCount { get; set; }
         /// <summary>
         /// 分拣数量
         /// </summary>
-        public int? Sorting { get; set; }
+        public double? Sorting { get; set; }
         /// <summary>
         /// 实际数量
         /// </summary>
-        public int? RealityCount { get; set; }
+        public double? RealityCount { get; set; }
 
         /// <summary>
         /// 不合格数量
